br still failing...

+++ src/Reach/Rendering/BatchRenderer.h
@@ -8,7 +8,7 @@
 #include <glm/vec4.hpp>
 #include <entt/entity/registry.hpp>
 #include <Reach/ECS/Components.h>
-
+#include <Reach/Tools/Input/InputManager.h>
 #define REACH_MAX_RENDERABLE 65536
 #define REACH_VERTEXDATA_SIZE sizeof(reach::VertexData)
 #define REACH_BUFFER_SIZE REACH_MAX_RENDERABLES * REACH_VERTEXDATA_SIZE
@@ -29,7 +29,7 @@ namespace reach{
         private:
             VertexData* _dataBuffer;
             uint32_t _vertexArrayID, _bufferID, _indexBufferID;
-            uint16_t _amountSubmitted;
+            uint32_t _amountSubmitted;
 
             void _setBuffer(VertexData data);
             uint16_t _ind[REACH_INDEX_BUFFER_SIZE];
